requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.4
django-cors-headers==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.8.0
flake8==3.5.0
greenlet==0.4.12
idna==2.6
jedi==0.11.0
logzero==1.3.1
mccabe==0.6.1
msgpack-python==0.4.8
mutagen==1.39
neovim==0.2.0
numpy==1.13.0
pafy==0.5.3.1
parso==0.1.0
pathlib==1.0.1
pbr==3.0.0
pep8==1.7.1
prettyconf==1.2.3
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.3
requests==2.18.4
scipy==0.19.1
six==1.11.0
spotipy==2.4.4
stevedore==1.21.0
titlecase==0.12.0
unicode-slugify==0.1.3
Unidecode==1.0.22
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
youtube-dl==2017.12.31
